ACTIVEMQ6-42 Fix FORCED_DELIVERY_MESSAGE dependent on hornetq naming
When dealing with receiveNoWait messages, a special message with
a property called _activemq.forced.delivery.seq is being sent.

This property used to be _hornetq.forced.delivery.seq. In order
to be compatible with 'old' hornetq clients, this fix does the
following:

1. Determine the client type of a client during protocol handshake
phase of a connection. New clients will connect with a 'ACTIVEMQ'
and old clients will do with either 'HORNETQ' or nothing.

2. The clien types are identified by a new enum type ClientType.
It has two values 'AMQ' (new clients) and 'HQ' (old ones).

3. When a forced delivery is being sent, put a proper property
to the special message based on the client type. If it is 'AMQ'
client, use _activemq.forced.delivery.seq, if it is 'HQ' use
_hornetq_forced.delivery.seq.